Initial Commit
This is an updated version of the AFNetworking Crash Course tutorial on
Ray Wenderlich's website. This is updated to include the new version of
AFNetworking (2.0.3). At the time of this commit the three links from
Ray's site are not producing data, so only the API button will be
displayed without errors.

http://www.raywenderlich.com/downloads/weather_sample/weather.php?format
=json
http://www.raywenderlich.com/downloads/weather_sample/weather.php?format
=xml
http://www.raywenderlich.com/downloads/weather_sample/weather.php?format
=plist

I assume the files will eventually be displayed correctly on Ray's
site, at which point this note can be disregarded.